Refactor monitoring
