+++ test/unit/libraries/mysterium-client/service-manager/service-manager-installer.spec.js
@@ -24,7 +24,8 @@ import SystemMock from '../../../../helpers/system-mock'
 import type { SystemMockManager } from '../../../../helpers/system-mock'
 import type { System } from '../../../../../src/libraries/mysterium-client/system'
 import ServiceManager from '../../../../../src/libraries/mysterium-client/service-manager/service-manager'
-import MonitoringMock from '../../../../helpers/mysterium-client/monitoring-mock'
+import { MockStatusNotifier } from '../../../../helpers/mysterium-client/monitoring-mock'
+import Monitoring from '../../../../../src/libraries/mysterium-client/monitoring/monitoring'
 
 const SERVICE_MANAGER_DIR = '/service-manager/bin'
 const SERVICE_MANAGER_PATH = SERVICE_MANAGER_DIR + '/servicemanager.exe'
@@ -87,8 +88,9 @@ describe('ServiceManagerInstaller', () => {
     const systemMock = createSystemMock()
     system = (systemMock: System)
     systemMockManager = (systemMock: SystemMockManager)
-    const monitoringMock = new MonitoringMock()
-    serviceManager = new ServiceManager(SERVICE_MANAGER_PATH, system, monitoringMock)
+    const mockNotifier = new MockStatusNotifier()
+    const monitoring = new Monitoring(mockNotifier)
+    serviceManager = new ServiceManager(SERVICE_MANAGER_PATH, system, monitoring)
   })
 
   describe('.needsInstallation()', () => {
